devcli: Delete duplicate users by PK/SK
We must only pass the unique keys to the deletion task, or
dynamodb complains by duplicates (even on a delete!)
